Rename outer iter module into iterator.
