+++ heed/src/iterator/range.rs
